requirements.txt
asarPy==1.0.1
Cython==3.1.1
keyboard==0.13.5
Mako>=1.3.0
Markdown==3.5.2
MarkupSafe==3.0.2
meson==1.5.2
packaging==25.0
pip==25.1.1
psycopg2-binary==2.9.10
pycairo==1.28.0
Pygments==2.19.1
PyGObject==3.50.1
setuptools==80.9.0
six==1.17.0
wheel==0.46.1
